+++ parley/src/main/java/nu/parley/android/data/model/ApiVersion.java
@@ -1,0 +1,16 @@
+package nu.parley.android.data.model;
+
+public enum ApiVersion {
+    /**
+     * @deprecated
+     * It is recommended to use {@link ApiVersion#V1_6} to support the latest functionality of Parley.
+     */
+    @Deprecated
+    V1_2,
+    /**
+     *
+     *
+     * This is the latest supported version by the library.
+     */
+    V1_6
+}